Move Wyre types to their own file
